gui/requirements.txt
fonttools
matplotlib
numpy
pyserial
